requirements.txt

venvdocutils==0.16
Markdown==3.2.1
mongoquery==1.3.6
objectpath==0.6.1
parso==0.6.2
pytest==5.4.1
pytest-json-report==1.2.1
PyYAML==5.3.1
typer==0.1.1
